+++ src/dotnet-gqlgen/ISchemaNode.cs
@@ -1,6 +1,0 @@
-namespace dotnet_gqlgen
-{
-    internal interface ISchemaNode
-    {
-    }
-}